refactor:#162 meal post add image 에서 버튼 합치기
